t/api.t: Remove microexception class use in flow control
And use local + passing values instead.

This was foolishly driven by the desire for all context about the
failure to automatically be in the "guard" block, so that the guard
could be accessed with $_.

But then I realised that existed prior to the ability to have a
lexical instead.

So instead, the `gsubtest` main block now is to setup lexicals
that are interesting to study, and the check block can either

a: Check the lexical directly
b: use any values in the guard via $_

Put anything in the guard that might be useful for diagnostics.

Keep things out that would be annoying and pass via lexicals.